sink/requirements.txt
aiokafka==0.10.0
psycopg2-binary==2.9.9